decouple notes
