+++ server/src/database/columns.go
@@ -49,7 +49,7 @@ type ColumnUpdate struct {
 // CreateColumn creates a new column. The index will be set to the highest available or the specified one. All other
 // indices will be adopted (increased by 1) to the new index.
 func (d *Database) CreateColumn(column ColumnInsert) (Column, error) {
-	maxIndexSelect := d.Db.NewSelect().Model((*Column)(nil)).ColumnExpr("COUNT(*) as index").Where("board = ?", column.Board)
+	maxIndexSelect := d.db.NewSelect().Model((*Column)(nil)).ColumnExpr("COUNT(*) as index").Where("board = ?", column.Board)
 
 	newIndex := math.MaxInt
 	if column.Index != nil {
@@ -60,9 +60,9 @@ func (d *Database) CreateColumn(column ColumnInsert) (Column, error) {
 		}
 	}
 
-	query := d.Db.NewInsert()
+	query := d.db.NewInsert()
 	if column.Index != nil {
-		indexUpdate := d.Db.NewUpdate().Model((*Column)(nil)).Set("index = index+1").Where("index >= ?", newIndex).Where("board = ?", column.Board)
+		indexUpdate := d.db.NewUpdate().Model((*Column)(nil)).Set("index = index+1").Where("index >= ?", newIndex).Where("board = ?", column.Board)
 		query = query.With("indexUpdate", indexUpdate)
 	}
 
@@ -84,9 +84,9 @@ func (d *Database) UpdateColumn(column ColumnUpdate) (Column, error) {
 		newIndex = 0
 	}
 
-	selectPrevious := d.Db.NewSelect().Model((*Column)(nil)).Column("board", "index").Where("id = ?", column.ID).Where("board = ?", column.Board)
-	maxIndexSelect := d.Db.NewSelect().Model((*Column)(nil)).Column("index").Where("board = ?", column.Board)
-	updateOnSmallerIndex := d.Db.NewUpdate().
+	selectPrevious := d.db.NewSelect().Model((*Column)(nil)).Column("board", "index").Where("id = ?", column.ID).Where("board = ?", column.Board)
+	maxIndexSelect := d.db.NewSelect().Model((*Column)(nil)).Column("index").Where("board = ?", column.Board)
+	updateOnSmallerIndex := d.db.NewUpdate().
 		Model((*Column)(nil)).
 		Column("index").
 		Set("index = index+1").
@@ -94,7 +94,7 @@ func (d *Database) UpdateColumn(column ColumnUpdate) (Column, error) {
 		Where("board = ?", column.Board).
 		Where("(SELECT index FROM \"selectPrevious\") > ?", newIndex).
 		Where("index >= ?", newIndex)
-	updateOnGreaterIndex := d.Db.NewUpdate().
+	updateOnGreaterIndex := d.db.NewUpdate().
 		Model((*Column)(nil)).
 		Column("index").
 		Set("index = index-1").
@@ -104,7 +104,7 @@ func (d *Database) UpdateColumn(column ColumnUpdate) (Column, error) {
 		Where("index <= ?", newIndex)
 
 	var c Column
-	_, err := d.Db.NewUpdate().
+	_, err := d.db.NewUpdate().
 		With("selectPrevious", selectPrevious).
 		With("maxIndexSelect", maxIndexSelect).
 		With("updateOnSmallerIndex", updateOnSmallerIndex).
@@ -121,17 +121,17 @@ func (d *Database) UpdateColumn(column ColumnUpdate) (Column, error) {
 // DeleteColumn deletes a column and adapts all indices of the other columns.
 func (d *Database) DeleteColumn(board, column, user uuid.UUID) error {
 	var columns []Column
-	selectPreviousIndex := d.Db.NewSelect().Model((*Column)(nil)).Column("index", "board").Where("id = ?", column)
-	indexUpdate := d.Db.NewUpdate().
+	selectPreviousIndex := d.db.NewSelect().Model((*Column)(nil)).Column("index", "board").Where("id = ?", column)
+	indexUpdate := d.db.NewUpdate().
 		With("selectPreviousIndex", selectPreviousIndex).
 		Model((*Column)(nil)).Set("index = index-1").
 		Where("board = (SELECT board from \"selectPreviousIndex\")").
 		Where("index >= (SELECT index from \"selectPreviousIndex\")")
-	boardUpdate := d.Db.NewUpdate().
+	boardUpdate := d.db.NewUpdate().
 		Model((*Board)(nil)).
 		Set("shared_note = null").
 		Where("id = ? AND (SELECT \"column\" FROM notes WHERE id = (SELECT shared_note FROM boards WHERE id = ?)) = ?", board, board, column)
-	_, err := d.Db.NewDelete().
+	_, err := d.db.NewDelete().
 		With("boardUpdate", boardUpdate).
 		With("indexUpdate", indexUpdate).
 		Model((*Column)(nil)).
@@ -145,13 +145,13 @@ func (d *Database) DeleteColumn(board, column, user uuid.UUID) error {
 // GetColumn returns the column for the specified id.
 func (d *Database) GetColumn(board, id uuid.UUID) (Column, error) {
 	var column Column
-	err := d.Db.NewSelect().Model(&column).Where("board = ?", board).Where("id = ?", id).Scan(context.Background())
+	err := d.db.NewSelect().Model(&column).Where("board = ?", board).Where("id = ?", id).Scan(context.Background())
 	return column, err
 }
 
 // GetColumns returns all columns for the specified board.
 func (d *Database) GetColumns(board uuid.UUID) ([]Column, error) {
 	var columns []Column
-	err := d.Db.NewSelect().Model(&columns).Where("board = ?", board).Order("index ASC").Scan(context.Background())
+	err := d.db.NewSelect().Model(&columns).Where("board = ?", board).Order("index ASC").Scan(context.Background())
 	return columns, err
 }
