+++ server/src/database/database_test.go
@@ -5,6 +5,8 @@ import (
 	"database/sql"
 	"fmt"
 	"log"
+	"scrumlr.io/server/notes"
+	"scrumlr.io/server/votes"
 	"time"
 
 	"github.com/ory/dockertest/v3"
@@ -19,6 +21,7 @@ import (
 )
 
 var testDb *Database
+var noteDB notes.NotesDatabase
 var reactionDb reactions.ReactionDatabase
 var fixture *dbfixture.Fixture
 
@@ -46,6 +49,7 @@ func testMainWithDefer(m *testing.M) int {
 
 	testDb = New(bun)
 	reactionDb = reactions.NewReactionsDatabase(bun)
+	noteDB = notes.NewNotesDatabase(bun)
 	err = loadTestdata()
 	if err != nil {
 		println(fmt.Sprintf("unable to load testdata: %s", err))
@@ -114,18 +118,18 @@ func initDatabase() (string, func(), error) {
 }
 
 func loadTestdata() error {
-	testDb.Db.RegisterModel(
+	testDb.db.RegisterModel(
 		(*User)(nil),
 		(*Board)(nil),
 		(*BoardSessionInsert)(nil),
 		(*Column)(nil),
-		(*Note)(nil),
-		(*Voting)(nil),
-		(*Vote)(nil),
+		(*notes.Note)(nil),
+		(*votes.Voting)(nil),
+		(*votes.Vote)(nil),
 		(*reactions.DatabaseReaction)(nil),
 		(*BoardTemplate)(nil),
 		(*ColumnTemplate)(nil),
 	)
-	fixture = dbfixture.New(testDb.Db)
+	fixture = dbfixture.New(testDb.db)
 	return fixture.Load(context.Background(), os.DirFS("testdata"), "fixture.yml")
 }
