+++ server/src/database/health.go
@@ -1,11 +1,11 @@
 package database
 
 import (
-  "context"
+	"context"
 )
 
 func (d *Database) IsHealthy() bool {
-  var result int
-  err := d.Db.NewSelect().Model((*Board)(nil)).ModelTableExpr("").ColumnExpr("1").Scan(context.Background(), &result)
-  return err == nil
+	var result int
+	err := d.db.NewSelect().Model((*Board)(nil)).ModelTableExpr("").ColumnExpr("1").Scan(context.Background(), &result)
+	return err == nil
 }
