+++ server/src/notes/dto.go
@@ -2,6 +2,8 @@ package notes
 
 import (
 	"github.com/google/uuid"
+	"github.com/uptrace/bun"
+	"time"
 )
 
 // NoteCreateRequest represents the request to create a new note.
@@ -44,6 +46,50 @@ type NoteDeleteRequest struct {
 
 	// Delete note or the complete stack.
 	DeleteStack bool `json:"deleteStack"`
+}
+
+type NoteDB struct {
+	bun.BaseModel `bun:"table:notes"`
+	ID            uuid.UUID
+	CreatedAt     time.Time
+	Author        uuid.UUID
+	Board         uuid.UUID
+	Column        uuid.UUID
+	Text          string
+	Stack         uuid.NullUUID
+	Rank          int
+	Edited        bool
+}
+
+type NoteInsertDB struct {
+	bun.BaseModel `bun:"table:notes"`
+	Author        uuid.UUID
+	Board         uuid.UUID
+	Column        uuid.UUID
+	Text          string
+}
+
+type NoteImportDB struct {
+	bun.BaseModel `bun:"table:notes"`
+	Author        uuid.UUID
+	Board         uuid.UUID
+	Text          string
+	Position      *NoteUpdatePosition `bun:",embed"`
+}
+
+type NoteUpdatePosition struct {
+	Column uuid.UUID
+	Rank   int
+	Stack  uuid.NullUUID
+}
+
+type NoteUpdateDB struct {
+	bun.BaseModel `bun:"table:notes"`
+	ID            uuid.UUID
+	Board         uuid.UUID
+	Text          *string
+	Position      *NoteUpdatePosition `bun:"embed"`
+	Edited        bool
 }
 
 type NoteSlice []*Note
