+++ server/src/services/services.go
@@ -3,15 +3,10 @@ package services
 import (
 	"context"
 
-	"scrumlr.io/server/columns"
-	"scrumlr.io/server/notes"
-	"scrumlr.io/server/votes"
-
 	"github.com/google/uuid"
 	"scrumlr.io/server/columns"
 	"scrumlr.io/server/common/dto"
 	"scrumlr.io/server/common/filter"
-	"scrumlr.io/server/notes"
 )
 
 type Users interface {
@@ -70,15 +65,6 @@ type BoardSessions interface {
 	ModeratorSessionExists(ctx context.Context, boardID, userID uuid.UUID) (bool, error)
 	SessionRequestExists(ctx context.Context, boardID, userID uuid.UUID) (bool, error)
 	ParticipantBanned(ctx context.Context, boardID, userID uuid.UUID) (bool, error)
-}
-
-type Notes interface {
-	Create(ctx context.Context, body dto.NoteCreateRequest) (*notes.Note, error)
-	Import(ctx context.Context, body dto.NoteImportRequest) (*notes.Note, error)
-	Get(ctx context.Context, id uuid.UUID) (*notes.Note, error)
-	Update(ctx context.Context, body dto.NoteUpdateRequest) (*notes.Note, error)
-	List(ctx context.Context, id uuid.UUID) ([]*notes.Note, error)
-	Delete(ctx context.Context, body dto.NoteDeleteRequest, id uuid.UUID) error
 }
 
 type Health interface {
