--- server/src/services/votings/votes.go
@@ -1,1 +1,0 @@
-package votings
